chore: Separate out `util` module into package
