+++ src/jobs/runner/kueue.py
@@ -5,7 +5,7 @@ from kubernetes import client
 from jobs import Image, Job
 from jobs.runner.base import Runner, _make_executor_command
 from jobs.types import K8sResourceKind
-from jobs.util import KubernetesNamespaceMixin, sanitize_rfc1123_domain_name
+from jobs.utils.kubernetes import KubernetesNamespaceMixin, sanitize_rfc1123_domain_name
 from jobs.utils.kueue import kueue_scheduling_labels
